v0.3.1
- Fixed `sharing as web page`. The 'sharing as webpage' feature now 
works again for both tab groups (share button) and all tabs (Share all 
button). However, the downside to this is that OneTab Reborn now uses 
the same key (specified in the manifest file) in order for the 'sharing 
as webpage' feature to work. Therefore, it is no longer possible to have 
both OneTab Reborn and OneTab installed at the same time.

- Made a few tweaks to the styles, made it easier to apply custom 
colours (if you want) in the CSS file by moving all the background 
colours to the top.

- Other under the hood changes and improvements.